Add files via upload
CTF for Hiring !  April, 2021

Thank you for applying to our Solidity Auditor role.  Your resume on paper looks good so we'd like to start the interview process with a practical assessment which will allow us to fully evaluate your skillset.

A Smart Contract is deployed on Ropsten. A 32 byte key in the contract's storage is used to read and write other 32 byte strings too a blockchain memory array.  
 
https://ropsten.etherscan.io/address/0x31dcfe9df827ba9065dd7f4057f68868eaacbdaf

Flag 1: What is the 32 Byte Key's Plain Text and Hex?
Flag 2: What is the first string added into blockchain memory? (First Index in Array)
Flag 3: Add your name "string" into the Array, and send us the what you uploaded along with the index number in memory where to find it.